Update `CHANGELOG.md` introducing new version `5.13.1`
I just really don't know that
I'm allowed to record this,
but the PR guide tells me to do it
at the Checklist,
so I'm doing it.
I think the patch version `x.x.1`
would be decent for this case
as I expect that this change would be
no harm to the plugin's main functionality.